rename arrays & movielens to private modules
