+++ lenskit/lenskit/testing/__init__.py
@@ -10,8 +10,8 @@ from contextlib import contextmanager
 
 import pytest
 
-from .arrays import coo_arrays, scored_lists, sparse_arrays, sparse_tensors
-from .movielens import demo_recs, ml_100k, ml_100k_zip, ml_ds, ml_ratings, ml_test_dir
+from ._arrays import coo_arrays, scored_lists, sparse_arrays, sparse_tensors
+from ._movielens import demo_recs, ml_100k, ml_100k_zip, ml_ds, ml_ratings, ml_test_dir
 
 __all__ = [
     "coo_arrays",
