+++ lenskit/lenskit/testing/_arrays.py
